backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
beautifulsoup4==4.12.2
billiard==4.2.1
celery==5.3.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
distlib==0.3.9
Django==4.2.7
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
executing==2.2.0
filelock==3.17.0
idna==3.10
ipython==8.32.0
jedi==0.19.2
kombu==5.5.4
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
Pillow==10.0.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pynput==1.8.1
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
pywifi==1.1.12
redis==5.0.1
requests==2.31.0
scapy==2.6.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
virtualenv==20.29.2
wcwidth==0.2.13
